+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Avro schema from file avsc
+- New alias for -t (--topic)
+- New alias for -b (--botstrap-servers)
+- New alias for -k (--key) and -v (--value)
+
+### Changed
+
+- Command --schema-registry change to --registry
+- Now the option -x becomes -c and --config
 
 ### Fixed